@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :rm_repo_folder, :class => RepositoryManager::RepoFolder do
     sequence :name do |n|
       'Folder'
