@@ -1,5 +1,5 @@
 class RepositoryManager::RepoFile < RepositoryManager::RepoItem
-  attr_accessible :file, :content_type, :file_size, :checksum if RepositoryManager.protected_attributes?
+  # attr_accessible :file, :content_type, :file_size, :checksum if RepositoryManager.protected_attributes?
 
   validates_presence_of :file
   mount_uploader :file, RepoFileUploader
@@ -164,11 +164,26 @@ class RepositoryManager::RepoFile < RepositoryManager::RepoItem
 
   private
 
+
+  def repo_file_params
+    params.require(:repo_file).permit(:file, :content_type, :file_size)
+  end
+
   def update_asset_attributes
     if file.present? && file_changed?
       self.content_type = file.file.content_type
       self.file_size = file.file.size
-      self.checksum = Digest::MD5.file(file.path).hexdigest
+      # self.checksum = Digest::MD5.file(file.path).hexdigest
+
+      # self.attributes = {
+      #   content_type: file.file.content_type ,
+      #   file_size: file.file.size ,
+      #   checksum:  Digest::MD5.file(file.path).hexdigest
+      #  }
+
+      # self.update_attribute(:content_type, file.file.content_type)
+      # self.update_attribute(:file_size, file.file.size)
+      # self.update_attribute(:checksum, Digest::MD5.file(file.path).hexdigest)
     end
   end
 
