@@ -1,10 +1,10 @@
 # encoding: utf-8
-require 'carrierwave/processing/mime_types'
+# require 'carrierwave/processing/mime_types'
 
 class RepoFileUploader < CarrierWave::Uploader::Base
-  include CarrierWave::MimeTypes
+  # include CarrierWave::MimeTypes
 
-  process :set_content_type
+  # process :set_content_type
 
   # Include RMagick or MiniMagick support:
   # include CarrierWave::RMagick
